modelView moved to Typescript and Browserify
