+++ src/modelView/model/index.ts
@@ -1,0 +1,7 @@
+/** 
+ * Module that defines the contract for communication 
+ * between the Extension and the plan view webview. 
+ */
+
+export * from './GraphViewData';
+export * from './stateViewModel';
