+++ views/modelView/static/graphView.html
@@ -4,16 +4,15 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <script src="../common/vis-network.min.js" type="text/javascript"></script>
-    <script src="../common/baseWebview.js" type="text/javascript"></script>
-    <script defer src="graphView.js" type="text/javascript"></script>
-    <script src="baseView.js" type="text/javascript"></script>
+    <script src="../../common/vis-network.min.js" type="text/javascript"></script>
+    <script src="../../common/baseWebview.js" type="text/javascript"></script>
+    <script defer src="../out/graphView.js" type="text/javascript"></script>
     <link rel="stylesheet" type="text/css" href="baseView.css">
     </link>
     <link rel="stylesheet" type="text/css" href="graphView.css">
     </link>
-    <link href="../common/vis-network.min.css" rel="stylesheet" type="text/css" />
-    <link rel="stylesheet" type="text/css" href="../common/codicon.css"></link>
+    <link href="../../common/vis-network.min.css" rel="stylesheet" type="text/css" />
+    <link rel="stylesheet" type="text/css" href="../../common/codicon.css"></link>
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <!-- CSP -->
     <title>AI Planning problem viewer for VS Code</title>
@@ -26,10 +25,10 @@
             <i class="codicon codicon-refresh" id="reset" title="Reset view to original layout"></i>
             <i class="codicon codicon-settings" id="toggleSettings" title="Show settings"></i>
             <i class="codicon codicon-move" id="fit" title="Fit all nodes" ></i>
-            <img src="img/light/directed-graph-top-down.svg" theme="light" id="topDown-light" title="Top down layout" class="menuButton" />
-            <img src="img/dark/directed-graph-top-down.svg" theme="dark" id="topDown-dark" title="Top down layout" class="menuButton" />
-            <img src="img/light/directed-graph-left-to-right.svg" theme="light" id="leftRight-light" title="Left-to-right layout" class="menuButton" />
-            <img src="img/dark/directed-graph-left-to-right.svg" theme="dark" id="leftRight-dark" title="Left-to-right layout" class="menuButton" />
+            <img src="../img/light/directed-graph-top-down.svg" theme="light" id="topDown-light" title="Top down layout" class="menuButton" />
+            <img src="../img/dark/directed-graph-top-down.svg" theme="dark" id="topDown-dark" title="Top down layout" class="menuButton" />
+            <img src="../img/light/directed-graph-left-to-right.svg" theme="light" id="leftRight-light" title="Left-to-right layout" class="menuButton" />
+            <img src="../img/dark/directed-graph-left-to-right.svg" theme="dark" id="leftRight-dark" title="Left-to-right layout" class="menuButton" />
         </span>
         <span id="insetMenu">
             <i class="codicon codicon-chevron-down" id="expandInset" title="Expand this view vertically"></i>
