+++ views/modelView/static/problemInitView.html
@@ -4,14 +4,13 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <script src="../common/vis-network.min.js" type="text/javascript"></script>
-    <script src="../common/baseWebview.js" type="text/javascript"></script>
-    <script defer src="problemInitView.js"></script>
-    <script src="baseView.js"></script>
+    <script src="../../common/vis-network.min.js" type="text/javascript"></script>
+    <script src="../../common/baseWebview.js" type="text/javascript"></script>
+    <script defer src="../out/problemInitView.js"></script>
     <link rel="stylesheet" type="text/css" href="baseView.css"></link>
     <link rel="stylesheet" type="text/css" href="problemInitView.css"></link>
-    <link rel="stylesheet" type="text/css" href="../common/codicon.css"></link>
-    <link href="../common/vis-network.min.css" rel="stylesheet" type="text/css" />
+    <link rel="stylesheet" type="text/css" href="../../common/codicon.css"></link>
+    <link href="../../common/vis-network.min.css" rel="stylesheet" type="text/css" />
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <!-- CSP -->
     <title>AI Planning problem viewer for VS Code</title>
