+++ views/modelView/test/test.ts
@@ -1,0 +1,10 @@
+import { expect } from 'chai';
+
+// test template
+describe("exampleClass", () => {
+	describe("exampleMethod", () => {
+		it('simple assertion should pass', () => {
+			expect(true).to.be.equal(true);
+		});
+	});
+});
